+++ packages/typespec-rust/src/codegen/mod.ts
@@ -8,18 +8,22 @@ import * as helpers from './helpers.js';
 import * as rust from '../codemodel/index.js';
 
 /**
- * emits the contents of the mod.rs file for clients
+ * emits the contents of the clients/mod.rs file
  * 
  * @param crate the crate for which to emit the mod.rs file
+ * @param addlMods any additional modules to include
  * @returns the contents of the mod.rs file
  */
 export function emitClientsModRs(crate: rust.Crate, addlMods: Array<string>): string {
   const content = helpers.contentPreamble();
   const body = new Array<string>();
 
+  const modules = new Array<string>();
   // first add the modules for each client
   for (const client of crate.clients) {
-    body.push(`mod ${codegen.deconstruct(client.name).join('_')};`);
+    const clientModule = codegen.deconstruct(client.name).join('_');
+    body.push(`mod ${clientModule};`);
+    modules.push(clientModule);
   }
 
   // add any additional mod entries
@@ -27,28 +31,20 @@ export function emitClientsModRs(crate: rust.Crate, addlMods: Array<string>): st
     body.push(`${addlMod};`);
   }
 
-  // now add re-exports for each client
-  for (const client of crate.clients) {
-    const clientModule = codegen.deconstruct(client.name).join('_');
-    if (client.constructable) {
-      body.push(`pub use ${clientModule}::*;`);
-      continue;
-    }
-
-    // for non-instantiable clients we only re-export method option types
-    const optionsTypes = new Array<string>();
-    for (const method of client.methods) {
-      if (method.kind !== 'clientaccessor') {
-        optionsTypes.push(method.options.type.name);
-      }
-    }
-    if (optionsTypes.length > 0) {
-      body.push(`pub use ${clientModule}::{${optionsTypes.join()}};`);
-    }
+  // now add re-exports for each client module
+  for (const module of modules) {
+    body.push(`pub use ${module}::*;`);
   }
+
   return content + body.join('\n');
 }
 
+/**
+ * emits the contents of the generated/mod.rs file
+ * 
+ * @param modules the modules to include
+ * @returns the contents of the mod.rs file
+ */
 export function emitGeneratedMod(modules: Array<string>): string {
   return helpers.contentPreamble() + modules.join(';\n') + ';\n';
 }