+++ packages/typespec-rust/test/spector/authentication/oauth2/src/lib.rs
@@ -5,4 +5,10 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
+
+pub use crate::generated::clients::{
+    OAuth2Client, OAuth2ClientInvalidOptions, OAuth2ClientOptions, OAuth2ClientValidOptions,
+};