+++ packages/typespec-rust/test/spector/azure/core/basic/src/lib.rs
@@ -5,8 +5,16 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
 
 pub mod models {
     pub use crate::generated::models::*;
 }
+
+pub use crate::generated::clients::{
+    BasicClient, BasicClientCreateOrReplaceOptions, BasicClientCreateOrUpdateOptions,
+    BasicClientDeleteOptions, BasicClientExportAllUsersOptions, BasicClientExportOptions,
+    BasicClientGetOptions, BasicClientListOptions, BasicClientOptions,
+};