+++ packages/typespec-rust/test/spector/azure/example/basic/src/lib.rs
@@ -5,9 +5,15 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
 
 pub mod models {
     pub use crate::generated::enums::*;
     pub use crate::generated::models::*;
 }
+
+pub use crate::generated::clients::{
+    BasicClient, BasicClientOptions, BasicServiceOperationGroupClientBasicOptions,
+};