+++ packages/typespec-rust/test/spector/azure/resource-manager/common-properties/src/lib.rs
@@ -5,9 +5,18 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
 
 pub mod models {
     pub use crate::generated::enums::*;
     pub use crate::generated::models::*;
 }
+
+pub use crate::generated::clients::{
+    CommonPropertiesClient, CommonPropertiesClientOptions,
+    CommonPropertiesManagedIdentityClientCreateWithSystemAssignedOptions,
+    CommonPropertiesManagedIdentityClientGetOptions,
+    CommonPropertiesManagedIdentityClientUpdateWithUserAssignedAndSystemAssignedOptions,
+};