+++ packages/typespec-rust/test/spector/client/structure/multi-client/src/lib.rs
@@ -5,8 +5,17 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
 
 pub mod models {
     pub use crate::generated::enums::*;
 }
+
+pub use crate::generated::clients::{
+    ClientAClient, ClientAClientOptions, ClientAClientRenamedFiveOptions,
+    ClientAClientRenamedOneOptions, ClientAClientRenamedThreeOptions, ClientBClient,
+    ClientBClientOptions, ClientBClientRenamedFourOptions, ClientBClientRenamedSixOptions,
+    ClientBClientRenamedTwoOptions,
+};