+++ packages/typespec-rust/test/spector/resiliency/srv-driven/new/src/lib.rs
@@ -5,4 +5,13 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
+
+pub use crate::generated::clients::{
+    ResiliencyServiceDrivenClient, ResiliencyServiceDrivenClientAddOperationOptions,
+    ResiliencyServiceDrivenClientFromNoneOptions,
+    ResiliencyServiceDrivenClientFromOneOptionalOptions,
+    ResiliencyServiceDrivenClientFromOneRequiredOptions, ResiliencyServiceDrivenClientOptions,
+};