+++ packages/typespec-rust/test/spector/server/endpoint/not-defined/src/lib.rs
@@ -5,4 +5,10 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
+
+pub use crate::generated::clients::{
+    NotDefinedClient, NotDefinedClientOptions, NotDefinedClientValidOptions,
+};