+++ packages/typespec-rust/test/spector/server/path/single/src/lib.rs
@@ -5,4 +5,8 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
+
+pub use crate::generated::clients::{SingleClient, SingleClientMyOpOptions, SingleClientOptions};