+++ packages/typespec-rust/test/spector/server/versions/not-versioned/src/lib.rs
@@ -5,4 +5,11 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
+
+pub use crate::generated::clients::{
+    NotVersionedClient, NotVersionedClientOptions, NotVersionedClientWithPathApiVersionOptions,
+    NotVersionedClientWithQueryApiVersionOptions, NotVersionedClientWithoutApiVersionOptions,
+};