+++ packages/typespec-rust/test/spector/type/model/empty/src/lib.rs
@@ -5,8 +5,15 @@
 
 mod generated;
 
-pub use crate::generated::clients::*;
+pub mod clients {
+    pub use crate::generated::clients::*;
+}
 
 pub mod models {
     pub use crate::generated::models::*;
 }
+
+pub use crate::generated::clients::{
+    EmptyClient, EmptyClientGetEmptyOptions, EmptyClientOptions,
+    EmptyClientPostRoundTripEmptyOptions, EmptyClientPutEmptyOptions,
+};